Sync PP

+++ wle_pp/wle_pp/src/pp/input/gamepad/gamepad_cores/gamepad_core.js
@@ -7,7 +7,20 @@ export class GamepadCore {
 
         this._myManagingHandPose = false;
 
+        this._myActive = true;
+        this._myStarted = false;
+
         this._myDestroyed = false;
+    }
+
+    setActive(active) {
+        this._setActiveHook(active);
+
+        this._myActive = active;
+    }
+
+    isActive() {
+        return this._myActive;
     }
 
     getHandedness() {
@@ -35,14 +48,23 @@ export class GamepadCore {
     }
 
     start() {
+        if (this._myStarted) return;
+
+        this._myStarted = true;
+
         if (this.getHandPose() && this.isManagingHandPose()) {
             this.getHandPose().start();
         }
 
         this._startHook();
+
+        this._myActive = false;
+        this.setActive(true);
     }
 
     preUpdate(dt) {
+        if (!this._myActive) return;
+
         if (this.getHandPose() && this.isManagingHandPose()) {
             this.getHandPose().update(dt);
         }
@@ -51,6 +73,8 @@ export class GamepadCore {
     }
 
     postUpdate(dt) {
+        if (!this._myActive) return;
+
         this._postUpdateHook(dt);
     }
 
@@ -68,6 +92,10 @@ export class GamepadCore {
     }
 
     // Hooks
+
+    _setActiveHook(active) {
+
+    }
 
     _startHook() {
 
@@ -88,7 +116,11 @@ export class GamepadCore {
     // Hooks end
 
     destroy() {
+        if (this._myDestroyed) return;
+
         this._myDestroyed = true;
+
+        this.setActive(false);
 
         this._destroyHook();
